move: VisitDate 클래스 및 menu 패키지 user 패키지로 이동
