+++ src/main/kotlin/christmas/domain/event/Gift.kt
@@ -1,6 +1,6 @@
 package christmas.domain.event
 
-import christmas.domain.menu.Drink
+import christmas.domain.user.menu.Drink
 
 enum class Gift(val menuName: String, val menuPrice: Int, val range: IntRange) {
     Champagne(Drink.Champagne.menuName, Drink.Champagne.price, 120_000..Int.MAX_VALUE);
