+++ src/test/kotlin/christmas/domain/event/DiscountTest.kt
@@ -1,9 +1,9 @@
 package christmas.domain.event
 
-import christmas.domain.VisitDate
-import christmas.domain.menu.Dessert
-import christmas.domain.menu.MainDish
-import christmas.domain.menu.Menu
+import christmas.domain.user.VisitDate
+import christmas.domain.user.menu.Dessert
+import christmas.domain.user.menu.MainDish
+import christmas.domain.user.menu.Menu
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
