+++ src/test/kotlin/christmas/domain/user/VisitDateTest.kt
@@ -1,5 +1,7 @@
-package christmas.domain
+package christmas.domain.user
 
+import christmas.domain.Error
+import christmas.domain.hasError
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.params.ParameterizedTest
